requirements.txt
openai
python-dotenv
fpdf  # ou reportlab, opcional
